requirements.txt
pdf2image==1.6.0
Pillow==6.1.0
PyPDF2==1.26.0
tk==0.1.0
